Make store capability traps value-dependent
As part of this, we need to honour MXR for capability loads for
consistency. This vastly simplifies the PTE logic and gives a bunch of
deduplication. Given it's now far closer to the base model, we should
look at providing the right hooks so we don't need to override the whole
PTE permissions checks.