Update to new trapping behaviour
As part of this, refactor the code a bit to avoid the duplication in
ext_ptw_error. Just like there is only one PTW_No_Permission for the
base model's page faults, with the exception code inferred from the
access, we only need the one enum value to indicate that this is a
capability page fault and can infer the rest.

See https://github.com/CTSRD-CHERI/cheri-architecture/pull/42.