Treat CCSeal with an out-of-bounds capability as a move
We currently make sure to handle authorities that went unrepresentable
due to a CCopyType, but not those that are still representable yet out
of bounds.

Whilst here, reorder the CCSeal checks to match CSeal more (it's a
series of branches all doing the same thing, so there's no functional
difference) and document both CSeal and CCSeal, hoisting the rather
length comments up to the top so as to not clutter the code so much
whilst also rendering them more nicely in cheri-architecture.

CSeal_core no longer makes much sense now the length check (which
happens quite late for CSeal) isn't shared.