Be sure to use correct register file for result of store conditional of capability.
Some SC instructions re-use the source register as the destination. To
ensure this works correctly for split register file implementations we
need to be careful about writing to the correct register file. This is
only a concern for store conditional of capabilities where the source
register is a capability register but the destination may be either a
capability or an general purpose register depending on the
instruction.