Fix CJAL jump target
PCC.address is not kept in sync with PC (except when jumping to a
capability), so we must use PC, mirroring the address computation in
AUIPCC. PC handling probably should be reworked so PC is an alias for
PCC.address, though there are other oddities with get_next_pc and
set_next_pc too (one uses an offset, one uses an address) so would be a
bit of an undertaking. Of course, removing the base offsetting would
also remove that issue.