Drop unused exception-based CInvoke exception definitions
These originate from the original CCall in CHERI-MIPS and do not apply
to CHERI-RISC-V where the only CInvoke is CHERI-MIPS's CCallFast (CCall
selector 1).

Note that CallTrap/ReturnTrap/TSSUnderflow are already marked as
reserved in the spec, but AccessCInvokeIDCViolation is still present
despite only applying to old versions of CHERI-MIPS where CCall
(selector 0) had a delay slot.