Remove support for split register files
Split register file support was never implemented and was removed from the
specification in https://github.com/CTSRD-CHERI/cheri-specification/pull/10.
This commit removes the tiny amount of support for split register files in
the model, which was just an always-false `haveSplitRegFile` and the
Clear instruction since that is CClear in the merged register file case.