Correctly handle c.clcsp with cd == 0
This instruction encoding is reserved (to match the c.cldsp/c.clqsp
encoding that we are using for capability loads/stores). Previously we
would decode this instruction as not valid for capmode and fall back to
the base model c.flwsp/c.fldsp decoding where destination register == 0 is
a valid encoding. Fix this by adding a new encdec clause to handle the
cd == 0 case.

Pointed out by @jrtc27