Disable PCC relocation by default
This is somewhat more invasive than the change to disable DDC relocation,
so we might want to remove the fallback code outright rather than keeping
a have_pcc_relocation() configuration option.
It also means that writing to SCRs that extend CSRs no longer needs to
check for unaligned bases and can update the address.