Update CHERI code to 0.9.3 spec
This includes many updates to the CHERI code to bring it into line with the 0.9.3 spec, and also updates to align with a more recent version of the upstream sail-riscv repo. Updates include:

* Capability levels
* CHERI PTE extensions
* Splitting c.modesw

It compiles but I have not tested the executable. Also this does not use the new CMake build system yet. My planned next steps are:

1. Merge this repo into `sail-riscv`, so we have a single repo. This will still use the `ext_` mechanism to integrate CHERI, but the CHERI code will be in `model/cheri` instead of `src`.
2. Switch to the CMake build system.
3. Add `extensionEnabled()`/`sys_enable_cheri()`.
4. De-duplicate the copy/pasted files, e.g. instead of `cheri_vmem_ptw.sail` replacing and modifying `riscv_vmem_ptw.sail`, have a single file with CHERI code that is disabled via `sys_enable_cheri()`.

Where possible we should try and keep the CHERI code in separate files still so it can be compiled-out easily. That is useful for the formal and SV backends.